+++ lib/src/constants/mp_constants.dart
@@ -31,7 +31,7 @@ const bool mpDebugAlwaysShowVersions = bool.fromEnvironment(
 
 // Debug constant to override the current version used when comparing to the
 // newest version available. If empty, the actual current version is used.
-const String mpDebugNewVersionInterfaceCurrentVersion = 'v0.2.35';
+const String mpDebugNewVersionInterfaceCurrentVersion = '';
 
 const String mpHelpPagePath = 'assets/help';
 const String mpHelpPageFlathubDisabled = 'flathub_disabled';
@@ -458,6 +458,8 @@ const String mpMapiahReleasesAPIURL =
 const String mpMapiahCommitsAPIURL =
     'https://api.github.com/repos/rsevero/mapiah/commits?per_page=$mpMapiahCommitsAPIPerPage';
 const String mpMapiahReleasesAPIHeaderAccept = 'application/vnd.github+json';
+const String mpHttpHeaderAcceptEncoding = 'Accept-Encoding';
+const String mpHttpHeaderAcceptEncodingGzip = 'gzip';
 const String mpHttpUserAgent = 'Mapiah';
 const String mpMapiahGithubReleasesURL =
     'https://github.com/rsevero/mapiah/releases/tag/';